fix: refactored apply files to independent apply folder
